web/requirements.txt
aniso8601>=9.0.1
cachetools>=4.2.4
certifi>=2021.10.8
charset-normalizer>=2.0.9
click>=8.0.3
docopt>=0.6.2
Flask>=2.0.2
Flask-Cors>=3.0.10
Flask-Login>=0.5.0
Flask-RESTful>=0.3.9
Flask-SQLAlchemy>=2.5.1
google-auth>=2.3.3
idna>=3.3
itsdangerous>=2.0.1
Jinja2>=3.0.3
MarkupSafe>=2.0.1
pip>=21.1.1
psycopg2-binary>=2.9.2
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pyOpenSSL>=21.0.0
python-dotenv>=0.19.2
pytz>=2021.3
requests>=2.26.0
rsa>=4.8
setuptools>=56.0.0
six>=1.16.0
SQLAlchemy>=1.4.27
SQLAlchemy-Utils>=0.37.9
urllib3>=1.26.7
Werkzeug>=2.0.2
yarg>=0.1.9